Enable security in task server. Added userinfo to request.

+++ task/src/main/java/com/heartiger/task/controller/TaskController.java
@@ -1,6 +1,9 @@
 package com.heartiger.task.controller;
 
+import static com.heartiger.task.constant.Constants.HEADER_USER_ID;
+
 import com.heartiger.task.converter.TaskForm2TaskInfoConverter;
+import com.heartiger.task.datamodel.CategoryInfo;
 import com.heartiger.task.datamodel.TaskInfo;
 import com.heartiger.task.dto.ResultDTO;
 import com.heartiger.task.enums.ResultEnum;
@@ -9,8 +12,8 @@ import com.heartiger.task.form.TaskForm;
 import com.heartiger.task.service.CategoryService;
 import com.heartiger.task.service.TaskService;
 import com.heartiger.task.utils.ResultDTOUtil;
+import javax.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
@@ -31,12 +34,18 @@ public class TaskController {
     }
 
     @GetMapping("/user/{id}")
-    public ResultDTO<Object> getAllTasks(@PathVariable int id){
+    public ResultDTO getAllTasks(@PathVariable int id, HttpServletRequest request){
+        if(notOwner(id, request))
+            return ResultDTOUtil.error(ResultEnum.TASK_PERMISSION_ERROR);
+
         return ResultDTOUtil.success(taskService.findAllTasksByOwnerId(id));
     }
 
     @GetMapping("/search")
-    public ResultDTO findTaskByIdAndUserId(@RequestParam int tid, @RequestParam int oid) {
+    public ResultDTO findTaskByIdAndUserId(@RequestParam int tid, @RequestParam int oid, HttpServletRequest request) {
+
+        if(notOwner(oid, request))
+            return ResultDTOUtil.error(ResultEnum.TASK_PERMISSION_ERROR);
         Optional<TaskInfo> result = taskService.findTaskInfoByIdAndUserId(tid, oid);
         if(result.isPresent()) {
             return ResultDTOUtil.success(result.get());
@@ -45,17 +54,27 @@ public class TaskController {
     }
 
     @PostMapping("/new")
-    public ResultDTO<Object> createTask(@Valid TaskForm taskForm, BindingResult bindingResult) {
-
+    public ResultDTO createTask(@Valid TaskForm taskForm, BindingResult bindingResult, HttpServletRequest request) {
         if(bindingResult.hasErrors())
             throw new TasksException(ResultEnum.PARAMS_ERROR.getCode(), bindingResult.getFieldError().getDefaultMessage());
+
+        if(notOwner(taskForm.getOwnerId(), request))
+            return ResultDTOUtil.error(ResultEnum.TASK_PERMISSION_ERROR);
+
+        Optional<CategoryInfo> categoryInfo = categoryService
+            .findCategoryByIdAndUserId(taskForm.getCategoryId(), taskForm.getOwnerId());
+        if(!categoryInfo.isPresent())
+            return ResultDTOUtil.error(ResultEnum.CATEGORY_ENTRY_NOT_FOUND);
 
         TaskInfo taskInfo = TaskForm2TaskInfoConverter.convert(taskForm);
         return ResultDTOUtil.success(taskService.saveTaskInfo(taskInfo));
     }
 
     @PostMapping("/complete")
-    public ResultDTO<Object> completeTask(@RequestParam int tid, @RequestParam int oid) {
+    public ResultDTO completeTask(@RequestParam int tid, @RequestParam int oid, HttpServletRequest request) {
+
+        if(notOwner(oid, request))
+            return ResultDTOUtil.error(ResultEnum.TASK_PERMISSION_ERROR);
 
         Optional<TaskInfo> result = taskService.findTaskInfoByIdAndUserId(tid, oid);
         if(!result.isPresent())
@@ -67,10 +86,14 @@ public class TaskController {
     }
 
     @PostMapping("/edit/{id}")
-    public ResultDTO editTask(@PathVariable int id, @Valid TaskForm taskForm, BindingResult bindingResult) {
+    public ResultDTO editTask(@PathVariable int id, @Valid TaskForm taskForm, BindingResult bindingResult, HttpServletRequest request) {
         if(bindingResult.hasErrors()) {
             throw new TasksException(ResultEnum.PARAMS_ERROR.getCode(), bindingResult.getFieldError().getDefaultMessage());
         }
+
+        if(notOwner(taskForm.getOwnerId(), request))
+            return ResultDTOUtil.error(ResultEnum.TASK_PERMISSION_ERROR);
+
         Optional<TaskInfo> taskToUpdate = taskService.findTaskInfoByIdAndUserId(id, taskForm.getOwnerId());
 
         if(!taskToUpdate.isPresent()) return new ResultDTO<>(ResultEnum.PARAMS_ERROR);
@@ -85,12 +108,21 @@ public class TaskController {
     }
 
     @DeleteMapping("/delete")
-    public ResultDTO deleteTask(@RequestParam int tid, @RequestParam int oid){
+    public ResultDTO deleteTask(@RequestParam int tid, @RequestParam int oid, HttpServletRequest request){
+
+        if(notOwner(oid, request))
+            return ResultDTOUtil.error(ResultEnum.TASK_PERMISSION_ERROR);
+
         Optional<TaskInfo> result = taskService.findTaskInfoByIdAndUserId(tid, oid);
         if(!result.isPresent())
             return ResultDTOUtil.error(ResultEnum.ENTRY_NOT_FOUND);
 
         taskService.deleteTaskByIdAndUserId(tid,oid);
         return ResultDTOUtil.success();
+    }
+
+    static boolean notOwner(@PathVariable int id, HttpServletRequest request) {
+        String userId = request.getHeader(HEADER_USER_ID);
+        return !String.valueOf(id).equals(userId);
     }
 }
